+++ src/ByteSync.ServerCommon/Commands/FileTransfers/GetDownloadFileUrlCommandHandler.cs
@@ -2,7 +2,7 @@ using ByteSync.ServerCommon.Interfaces.Services;
 using MediatR;
 using Microsoft.Extensions.Logging;
 
-namespace ByteSync.ServerCommon.Commands.FileTransfer;
+namespace ByteSync.ServerCommon.Commands.FileTransfers;
 
 public class GetDownloadFileUrlCommandHandler : IRequestHandler<GetDownloadFileUrlRequest, string>
 {
