+++ tests/ByteSync.ServerCommon.Tests/Commands/FileTransfers/AssertDownloadIsFinishedCommandHandlerTests.cs
@@ -1,69 +1,58 @@
 using ByteSync.Common.Business.SharedFiles;
 using ByteSync.ServerCommon.Business.Auth;
 using ByteSync.ServerCommon.Business.Sessions;
-using ByteSync.ServerCommon.Commands.FileTransfer;
+using ByteSync.ServerCommon.Commands.FileTransfers;
 using ByteSync.ServerCommon.Interfaces.Repositories;
 using ByteSync.ServerCommon.Interfaces.Services;
 using FakeItEasy;
 using FluentAssertions;
 using Microsoft.Extensions.Logging;
 
-namespace ByteSync.ServerCommon.Tests.Commands.FileTransfer;
+namespace ByteSync.ServerCommon.Tests.Commands.FileTransfers;
 
 [TestFixture]
-public class AssertFilePartIsDownloadedCommandHandlerTests
+public class AssertDownloadIsFinishedCommandHandlerTests
 {
     private ICloudSessionsRepository _mockCloudSessionsRepository;
-    private ISharedFilesService _mockSharedFilesService;
-    private ILogger<AssertFilePartIsDownloadedCommandHandler> _mockLogger;
-    private AssertFilePartIsDownloadedCommandHandler _assertFilePartIsDownloadedCommandHandler;
+    private ISynchronizationService _mockSynchronizationService;
+    private ILogger<AssertDownloadIsFinishedCommandHandler> _mockLogger;
+    private AssertDownloadIsFinishedCommandHandler _assertDownloadIsFinishedCommandHandler;
     private ITransferLocationService _mockTransferLocationService;
     
     [SetUp]
     public void Setup()
     {
         _mockCloudSessionsRepository = A.Fake<ICloudSessionsRepository>();
-        _mockSharedFilesService = A.Fake<ISharedFilesService>();
-        _mockLogger = A.Fake<ILogger<AssertFilePartIsDownloadedCommandHandler>>();
+        _mockSynchronizationService = A.Fake<ISynchronizationService>();
+        _mockLogger = A.Fake<ILogger<AssertDownloadIsFinishedCommandHandler>>();
         _mockTransferLocationService = A.Fake<ITransferLocationService>();
-        
-        _assertFilePartIsDownloadedCommandHandler = new AssertFilePartIsDownloadedCommandHandler(
+
+        _assertDownloadIsFinishedCommandHandler = new AssertDownloadIsFinishedCommandHandler(
             _mockCloudSessionsRepository,
-            _mockSharedFilesService,
+            _mockSynchronizationService,
             _mockLogger,
             _mockTransferLocationService);
     }
 
     [Test]
-    public async Task Handle_ValidRequest_AssertsFilePartIsDownloaded()
+    public async Task Handle_ValidRequest_AssertsDownloadIsFinished()
     {
         // Arrange
         var sessionId = "session1";
         var client = new Client { ClientInstanceId = "client1" };
         var sharedFileDefinition = new SharedFileDefinition { Id = "file1" };
-        var partNumber = 1;
 
-        var request = new AssertFilePartIsDownloadedRequest(sessionId, client, sharedFileDefinition, partNumber);
+        var request = new AssertDownloadIsFinishedRequest(sessionId, client, sharedFileDefinition);
 
         // Mock the session repository to return a valid session member
-        var mockSessionMember = new SessionMemberData { ClientInstanceId = client.ClientInstanceId };
+        var mockSessionMember = A.Fake<SessionMemberData>();
         A.CallTo(() => _mockCloudSessionsRepository.GetSessionMember(sessionId, client)).Returns(mockSessionMember);
-        
-        // Mock the transfer location service to return true for IsSharedFileDefinitionAllowed
-        A.CallTo(() => _mockTransferLocationService.IsSharedFileDefinitionAllowed(mockSessionMember, sharedFileDefinition))
-            .Returns(true);
-
-        // Mock the shared files service
-        A.CallTo(() => _mockSharedFilesService.AssertFilePartIsDownloaded(sharedFileDefinition, client, partNumber))
-            .Returns(Task.CompletedTask);
 
         // Act
-        await _assertFilePartIsDownloadedCommandHandler.Handle(request, CancellationToken.None);
+        await _assertDownloadIsFinishedCommandHandler.Handle(request, CancellationToken.None);
 
         // Assert
         A.CallTo(() => _mockCloudSessionsRepository.GetSessionMember(sessionId, client))
-            .MustHaveHappenedOnceExactly();
-        A.CallTo(() => _mockTransferLocationService.IsSharedFileDefinitionAllowed(mockSessionMember, sharedFileDefinition))
             .MustHaveHappenedOnceExactly();
     }
 
@@ -74,10 +63,9 @@ public class AssertFilePartIsDownloadedCommandHandlerTests
         var sessionId = "session1";
         var client = new Client { ClientInstanceId = "client1" };
         var sharedFileDefinition = new SharedFileDefinition { Id = "file1" };
-        var partNumber = 1;
         var expectedException = new InvalidOperationException("Test exception");
 
-        var request = new AssertFilePartIsDownloadedRequest(sessionId, client, sharedFileDefinition, partNumber);
+        var request = new AssertDownloadIsFinishedRequest(sessionId, client, sharedFileDefinition);
 
         // Mock the session repository to throw an exception
         A.CallTo(() => _mockCloudSessionsRepository.GetSessionMember(sessionId, client))
@@ -85,7 +73,7 @@ public class AssertFilePartIsDownloadedCommandHandlerTests
 
         // Act & Assert
         var exception = await FluentActions.Awaiting(() => 
-            _assertFilePartIsDownloadedCommandHandler.Handle(request, CancellationToken.None))
+            _assertDownloadIsFinishedCommandHandler.Handle(request, CancellationToken.None))
             .Should().ThrowAsync<InvalidOperationException>();
 
         exception.Which.Should().Be(expectedException);
